+++ ProfilesMicroService.Application/DTO/Receptionist/ReceptionistDTO.cs
@@ -1,13 +1,12 @@
-﻿namespace ProfilesMicroService.Application.Services.DTO
+﻿namespace ProfilesMicroService.Application.DTO.Receptionist
 {
-    public class PatientDTO
+    public class ReceptionistDTO
     {
         public string Id { get; set; }
         public string AccountId { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string MiddleName { get; set; }
-        public DateTime DateOfBirth { get; set; }
-        public string AccountPhoneNumber { get; set; }
+        public string OfficeId { get; set; }
     }
 }
