+++ ProfilesMicroService.Application/Services/CQRS/Queries/GetPatientWithoutAccountQuery.cs
@@ -1,5 +1,5 @@
 ﻿using MediatR;
-using ProfilesMicroService.Application.Services.DTO;
+using ProfilesMicroService.Application.DTO.Patient;
 using ProfilesMicroService.Domain.Entities.Models;
 
 namespace ProfilesMicroService.Application.Services.CQRS.Queries
